Reference for ProToPortal.jl

ProToPortal.authenticate

ProToPortal.authenticated

ProToPortal.autofix_code

ProToPortal.build_lazy_aicall

ProToPortal.conversation2transcript

ProToPortal.deauthenticate

ProToPortal.evaluate_code

ProToPortal.flash

ProToPortal.flash

ProToPortal.flash_has_message

ProToPortal.label_conversation

ProToPortal.load_conversations_from_dir

ProToPortal.parse_critic

ProToPortal.prepare_conversation

ProToPortal.render_messages

ProToPortal.render_template_messages

ProToPortal.send_to_model

ProToPortal.send_to_model

ProToPortal.template_variables

ProToPortal.update_message

ProToPortal.update_message!

ProToPortal.authenticate — Method

Stores the user id on the session.

ProToPortal.authenticated — Method

is_authenticated(session) :: Bool
is_authenticated(params::Dict{Symbol,Any}) :: Bool

Returns true if a user id is stored on the session.

ProToPortal.autofix_code — Method

Runs one iteration of airetry!

ProToPortal.build_lazy_aicall — Method

Constructs AIGenerate call that mimics if it was just executed

ProToPortal.conversation2transcript — Method

Flattens the full conversation for critic to use

ProToPortal.deauthenticate — Method

deauthenticate(session)
deauthenticate(params::Dict{Symbol,Any})

Removes the user id from the session.

ProToPortal.evaluate_code — Method

Code evaluator. Returns the evaluted code block (AICode) and the feedback string.

ProToPortal.flash — Method

flash(value::Any) :: Nothing

Stores value onto the flash.

ProToPortal.flash — Method

flash()

Returns the flash dict object associated with the current HTTP request.

ProToPortal.flash_has_message — Method

flash_has_message() :: Bool

Checks if there's any value on the flash storage

ProToPortal.label_conversation — Method

Labels the conversation based on the transcript

ProToPortal.load_conversations_from_dir — Method

Loads all conversations from a directory (or its sub-directories)

ProToPortal.parse_critic — Method

Parses critic's response to get the suggestions and whether they requested to stop

ProToPortal.prepare_conversation — Method

Prepares the conversation for sending to the LLM

ProToPortal.render_messages — Method

Returns rendered messages, check if messages need updating as we might have changed the display

ProToPortal.render_template_messages — Method

Renders template displayed messages

ProToPortal.send_to_model — Method

Sends the conversation to the LLM.

ProToPortal.send_to_model — Method

Sends the conversation to the Auto-Critic Template for evaluation and suggestions

ProToPortal.template_variables — Method

Extracts the variables from the template name (if found)

ProToPortal.update_message! — Method

Updates the message in the display, if needed

ProToPortal.update_message — Method

Updates the provided message with the new content (creates a new object)